Allow clearing the organisation filter on the users page
Turn off autoselect for the organisation filter, so the first matching
organisation isn't automatically selected when typing. This allows
clearing the text in the input.

However, the selected value will still be the ID of the most recently
selected organisation. But the input will also send through a query
parameter named "organisation_id_raw" which contains the text in the
input. If this is empty, the controller will clear the organisation_id
so that the search will include all organisations.