Add NotFoundError and register with error middleware
Add a NotFoundError class to raise when we want to render a 404 page.
Register this with the error handling middleware in application.rb so
that it will be handled at a 404 error and call the not_found
controller action in ErrorsController to render the 404 page.

We can replace wherever we directly render the 404 page in controllers
to instead raise a NotFoundError to make error handling more
consistent.