Handle a form in the wrong state when visiting /live or /archived
When visiting /live or /live/pages routes, check that the form is in
a live state. If it has been archived, redirect to the equivalent
page for an archived form. If the form has not been made live, render
the 404 page.

Do the equivalent for /archived routes, redirecting to /live if the
form is live, and rendering the 404 page if the form has never been
made live.