Remove unnecessary rescue_from calls in controllers
The rescue_from in users_controller is unnecessary as
ActiveRecord::RecordNotFound is already handled by the error handling
middleware to render the 404 page as it is in
config.action_dispatch.rescue_responses by default.

Remove the rescue_from in web_controller and instead register
ActiveResource::ResourceNotFound in config/application.rb to be
handled by the error handling middleware as well.