Add tasks to update form submission type
We had tasks to update the submission type in forms-api, but now forms
exist and are managed in forms-admin so we need to add these tasks
here.

These tasks do not update forms-api as forms-runner has been switched
to get form documents from forms-admin rather than forms-api and we
are imminently going to stop updating forms-api from forms-admin.