Update successful deletion banner messages for style
This updates the success banner messages for deleting a group, a form and a question to make them more consistent with our other success banner messages. 

This will also aid clarity for people when they delete the draft of a form that has a live version - as it reinforces that it is the draft of the form that has been deleted.