Add language column to FormDocument
To support retrieving forms in multiple languages, we add a language column
to the FormDocument table.

This column will be used when retrieving FormDocuments from the API.
A default value of "en" is set, which covers all existing forms, until
we add features for creating Welsh forms.

FormDocuments will be fetched using form_id, tag and language.

There can be only one FormDocument for a given form_id, tag and language.

For example for a form with id 1, tag "live" there can only be one FormDocument.

A new index is added to the FormDocument table to ensure that this constraint
is enforced.

The old index is removed.

Because adding an in index blocks writes in Postgres, it's added concurrently.

This might be overkill for our use case but it's recommended by the strong
migration readme which has been useful in the past.

https://github.com/ankane/strong_migrations?tab=readme-ov-file#good-8