Add form#save_draft! method
Add a new Forrm#save_draft! method, which saves the form and updates
the form status to live_with_draft or archived_with_draft if the state
is currently live or archived respectively.

This method will replace calls to FormRepository.save! and will be
called wherever the form is saved through the admin interface. We
might not want the state of a live/archived form to be updated when a
form is updated when we are updating Forms as part of a data
migration, which is why we're adding a method for this rather than
adding an `after_update` callback to update the Form state.

The method returns `true`, as places where FormRepository.save! is
called rely on a truthy return value to indicate success. However,
the existing behaviour of FormRepository.save! was to raise an
exception if the save failed so this has been replicated.