Remove PageRepository.destroy method
Replace with a direct call to Page#destroy_and_update_form!. The
handling for the deletion being unsuccessful due to a race condition
and the page alreday having been deleted has been removed. This was
handled by attempting to redirect to the question page and displaying
an error message. However, this would have resulted in a 404 being
rendered as the page no longer exists. Instead, we now allow the error
handling middleware to show a 404 page when an
ActiveRecord::RecordNotFound error is raised.