Remove ConditionRepository.destroy method
Replace with direct ActiveRecord calls. Call Condition#destroy when
we check the return value. Call Condition#destroy! when we don't to
make sure we don't show a success page if there was an error deleting
a condition.

This was the last method in ConditionRepository, so the file has been
deleted.