Remove ConditionRepository.destroy method
Replace with direct calls to Condition#destroy_and_update_form!.
Previously, we were catching ActiveRecord::RecordNotFound exceptions
raised by Condition#destroy!, and attempting to re-render the deletion
confirmation page. However, this would result in a 404 upon redirect
as the condition no longer exists. Instead, allow all exceptions to
propogate and the ActiveRecord:RecordNotFound error will be handled by
the error handling middleware to render the 404 page.

This was the last method in ConditionRepository, so the file has been
deleted.