Fix failing test for Condition#exit_page?
We use Condition#exit_page_markdown to check if the condition is an exit
page.

In our tests setup we set exit_page_markdown to "" and check the result
of #exit_page?.

Before adding the mobility gem, this test passed.

Adding mobility translations causes the test to fail because "" is
normalised to nil.

The translation is set to presence: false so that empty strings are
preserved.

We might need to be careful about similar issues in the future. We may
also want to explicitly check the English version of the translation to
ensure the English and welsh versions are consistent.