Change how form_slug is set for translations
We set the form_slug of form when the name is set.

Form_slug and name are translated, and form_slug must be correctly set
for the current locale.

Previously we were setting [:form_slug] directly, which meant that
translations were not being set correctly.

This commit changes the way we set form_slug, to use the mobility aware
method.