Add DataStruct to answer_settings for translations
In the Page model, the jsonb answer_settings column is used to store
settings for the page, such as the selection options for a selection
page.

The page model uses the DataStructType when loading and saving the
answers from the database.

Answer_settings is part of the translations for the page, so we need
to make the same change to the answer_settings attribute for the
translations.

Extending from the Mobility dynamically creates a new model class for
the translations.

https://www.rubydoc.info/gems/mobility/Mobility/Backends/Table

We open the translation model class and add the answer_settings
attribute after this.