Remove case for form without group in delete
When deleting a draft, we check if the form is in a group to decide
where to redirect the user to.

This is left over from before groups were introduced. It's now no longer
possible for a form to be created outside of a group.

Users should always be redirected to the group page when deleting a
draft.