Add details to group#show for managing group
Show a details component which explains how to manage a group on the
group#show page.

It contains different text depending on whether the current user is an
organisation admin or a standard user. And different text for an active
or a trial group.

If the organisation hasn't signed the MOU or has no admin users, the
details component is not rendered.