Add translations for new submission types
We are not releasing the ability to turn on JSON attachments for
submission emails to admin users yet. However, we'll be turning it on
in the backend for some forms to test it out.

This commit adds the necessary translations for the new submission
types that were added to support enabling JSON attachments so that the
live form page doesn't look broken when the feature is enabled for a
form.

The existing wording for when a CSV attachment is enabled has been
updated inline with the proposed wording for when we release the
ability to turn on JSON attachments for submission emails to admin
users.