Add service for reordering pages
Add a service with a method to generate a new page order when given an
array of pages and their desired new positions. This will attempt to
place all pages that have been given a new position into that position,
and move pages that have not been given a new postion up or down to
accommodate them.

This also accounts for the case where multiple pages are given the
same position.

We do not account for pages being given positions outside the range of
the number of pages, as this should be handled by validation for the
input model.

Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>