Change how form_slug is set for translations
We set the form_slug of form when the name is set.

Form_name is translated, but form_slug stays in English regardless of
locale.

Previously we were setting [:form_slug] directly, which meant that
form_slug was being set to the Welsh text for both English and Welsh if
the Welsh name of the form changed.

This commit changes the way we set form_slug, to always use the English
name to generate the slug.