Add method to update page order
Add a method to the ReorderPagesService that accepts an array of page
IDs and their new positions.

The page that we will build to do the reordering will include a button
to preview the new order. If this has been clicked prior to the page
being submitted, we expect to have already called
`ReorderPagesService.generate_new_page_order` to order the pages. It
will also be possible to submit the page without previewing, so we
need to call `.generate_new_page_order` within the update method to
ensure the new order has been applied.

We are using the `acts_as_list_no_update` method provided by the
`acts_as_list` gem to update the positions of the page without the
`acts_as_list` callbacks being triggered for each individual update.

If a page has been added to the form that does not exist in the
provided array of page IDs, we raise an error. We will handle this by
showing a validation error to the user.

Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>