Add s3_with_json submission type
Add an additional allowed value for `submission_type` of
`s3_with_json`. If the submission_type has this value, we will send a
JSON file to the S3 bucket instead of a CSV file.

We will do some later work to separate the submission type s3/email
from the format csv/json, but for now this is a simple way to add the
functionality we need.

Update the rake task for configuring S3 submissions to require a
format parameter which is either "csv" or "json".