Move method to update form for a page to page model
When a page is updated, we want to set the question_section_completed
property on the form to false, and update the form document.

For changing the page order, we only want to update the form once
rather than for each page that is updated so that we don't need to
regenerate the form document multiple times. To achieve this, add a
method to the Form model to update the question_section_completed
property to false and save the form so we can call it from the
ChangeOrderService without duplicating code.