Update WelshTranslationInput to change attributes
Change the WelshTranslationInput so that it can change the welsh translations
of a form.

In this commit we add attributes to match the translated fields of form.

We also add mark_complete to the attributes. Without this, we'll break
the mark complete functionality.

Helpers are added to decide which attributes should be changed. They
look at the form's default values to decide if a feature is used.

These aren't directly tested. They are tested indirectly by the
Welsh submit and assign form values tests.