Add Welsh translation form attributes
Add attributes for the welsh translation form to the view.

This commit shouldn't be shown to the public until we've done the
styling and checked accessibility.

The fields are styled using tables and the labels are hidden from
displaying, except for screen readers. This is only for internal use.
I'm not sure if this is the best way to do this.

There is also a lot of duplication in the markup. We should be able to
DRY this up when we've got all the fields in place.

We conditionally show some fields depending on whether the form uses
them.

Because a form filler can access this page at any point while creating a
form we need to be careful about showing some fields.

Declaration is an optional field and the designs include a message to
show when a form isn't using it.

Payment URL is also an optional field but the designs don't include a
message, it's just not shown.

Support information is harder to show to if they haven't completed that
section yet. We wouldn't know which of the contact details to show.

Instead we just hide the whole section if they haven't completed it yet.

This might change.