Add custom field IDs to avoid clash
By default, the govuk form builder gem will generate the same id for all
fields with the same form name and attribute. This means that when using
fields_for, multiple fields with the same id can be generated. This is
invalid HTML, and also can cause accessibility issues as it's ambiguous
which fields the labels for the form fields are meant to point at.

This commit customises the id of each field and its label, so that each
form field will have a unique id.