Add submission format to forms
Add a string array column for submission format(s). This will eventually
hold the format of the submissions (i.e. "csv", "json", or any
combination), but for now it is not used anywhere.

We add a commented out enum declaration to the model code for
documentation. Given that we don't do any validation on the submission
type enum currently, we're not losing anything by not having the enum
properly defined.

We don't set a default because we want to be able to tell whether the
submission format has been assigned and is empty, or if it has not been
assigned. We could decide to add a default of an empty array in future.