Change ApplicationLogger to work with more than just string or hash
I was getting a build error from Vite locally, but because Vite Ruby
logs the process status the error message was being obscured by the
error from the application logger, which meant I had no idea what the
problem was without using the debugger.

Instead just coercing the message to a string if it's not a hash should
work for almost all circumstances, and now I can easily see what the
error is from Vite.