Fix assigning blank form name
Currently doing trying to set the form name to an empty string results
in the following error:

```
forms-admin(dev)> form.name = ""
app/models/form.rb:84:in 'Form#name=': undefined method 'parameterize' for nil (NoMethodError)

    self[:form_slug] = name_en.parameterize
                              ^^^^^^^^^^^^^
        from (forms-admin):2:in '<main>'

```

Instead we expect the assignment to succeed and then there to be a
validation error when the form is saved.

This is only something we ever want to do in tests, and this change
doesn't affect users as we do validation in the input object separately,
but I think it's still worth fixing.