Move code that changes forms from form state machine to form model
Refactors FormStateMachine so the bits that change attributes on the
form record itself are in the form model rather than in the state
machine.

This has the downside of requiring a bit more reading to figure out what
exactly happens to the form object on each event, but is easier to test
and makes it easier to read each file in isolation. It also means we no
longer need to make the `FakeForm` inherit from `Form` in a slightly
weird way.

It also turns out we can use the name of the method as a symbol instead
of wrapping the invocation in a proc, which seems nicer to me.