Refactor saving a draft of a form
This commit changes how saving a draft with `save_draft!` works so that
the form is saved through ActiveRecord only once. The main advantage of
this is then the [ActiveRecord::Dirty] module works as expected, and the
`previous_changes` of the form can be inspected after calling
`save_draft!` to see what changes were actually saved. This also reduces
some of the database calls needed.

As part of this change we also add `whiny_persistence` to true in the
state machine, so that the bang methods behave in the same way as the
ActiveRecord::Persistence bang methods, and raise an exception if the
form is invalid.

[ActiveRecord::Dirty]: https://api.rubyonrails.org/classes/ActiveModel/Dirty.html