Remove 'small' attribute from checkboxes
We were configuring our checkboxes with `small: true`. There are a few
problems with this:

- the `govuk_design_system_formbuilder` gem expects the `small: true` to
be on the `govuk_check_boxes_fieldset`, not the `govuk_check_box`
itself. As a result, instead of rendering checkboxes with the small
styles, it just renders checkboxes with a `small="true"` HTML attribute,
which isn't valid HTML and doesn't do anything.
- we don't actually want these checkboxes to be small - using large
checkboxes here is consistent with the design system guidance, our own
designs, other checkoboxes on the site, and the way these checkboxes
have been appearing since we first implemented them.

This commit removes the `small: true` to resolve these issues.