Add method to generate unique external ID
Whilst it should be extremely unlikely that we generate a duplicate
external ID for a group it is not impossible. If this happened, an
error would be raised creating a group and the user would see a
server error page.

We probably don't need to mitigate this for groups, but we will be
adding external IDs to pages, and would prefer to keep the external ID
short so that it doesn't make it too hard to read meaningful parts of
the URL. There will be a higher chance of collision for pages as we
expect a higher volume of pages in the future, potentially reaching
the hundreds of thousands or millions across the platform.

We're very unlikely to see a collision in the near future, but adding
this mitigation adds a single low-cost database query so should not
negatively impact performance in any meaningful way.