Update form documents to use Page external IDs
When converting a Page into a step hash for form document generation,
use the Page's external_id instead of its database ID.

At the same time, also add in a temporary "database_id" attribute
field set to the page's "id". Forms-runner will use this field so that
it is able to retrieve existing answers from in-progress user sessions
that were stored using the database ID as the key. We can remove this
field in a future migration once we can be pretty sure all in-progress
sessions have been completed or expired.

We will add a subsequent data migration to update existing form
documents to have this change.

Users shouldn't experience any noticeable interruption when the ID is
updated, as if they visit a URL containing the old database ID, they
will be redirected to the first un-completed question.