Make reverting a draft form work with external IDs
Make reverting a draft form work when the form document steps use
page external IDs instead of database IDs.

To avoid having to make this backwards compatible until the data
migration has been run to update all form documents to use internal
page IDs, raise an error if the migration hasn't been run yet.

Reverting the draft of a form is not a regularly used feature, so
breaking this temporarily is acceptable.