Add one off data migration to add external_id to pages
For existing pages, we want to generate an external_id. This rake task
assigns a random unique external_id to any Page records that do not
already have one. It then updates the steps in all FormDocuments to
have the same external_id as the associated Page.

If a Page for a step in a form document does not exist in the database
because it has been deleted from the draft form, generate a new random
external_id for that step.