Update task status to look at submission_format
Update the task list status for the task to configure submission email
attachments to look at the `submission_format` for the form to determine
whether any submission formats have been enabled.

If the `submission_type` is "s3", return optional. We probably don't
want to include the task at all so the status shouldn't be relevant in
this case.