Add submission attachments controller and view
Adds a submission attachments controller, where a form submission_format
can be set to blank, csv or json.

If the json_submission_enabled feature flag is not turned on, it will
return a 404 when accessing the controller's actions.

A success message is played if the form's submission_format has been
changed - depending on what it has been changed to.