Refactor revert_draft_form_service_spec to show better errors
Unfortunately, the custom matcher was making it hard to see what the
difference between the expected and actual values were when the test
failed.

Instead of using a custom matcher, use a method containing shared
expectations to reduce duplication while still allowing RSpec to show
the full diff between the form document content hashes when the tests
fail.