Add selection options to Welsh translation
To allow testing routes, we allow the welsh translation to have selection options.

This takes the simplest approach, presenting the selection options as
text fields under the page fields.

We need to take care when doing this to make sure the Welsh translation
have answer settings which match the English version.

The only text in any page's answer settings which needs to be translated
are the selection options. When assigning attributes, we check to see if
the Welsh has answer_settings. If it doesn't we copy the English
answer_settings and then reset the selection options to be blank.

When saving the translation, we need to copy the selection options from
the Welsh answer_settings to the English answer_settings.

This code is hacky and shouldn't be used in production. We need to take
more care about keeping them in sync and consider how to handle new
options being added before the translations are saved.

The view itself doesn't use the values directly from the Welsh
answer_settings and currently it relies on position, which I think is
wrong.

We modify the Page model to add value to the selection options. We use
this as a backstop to add the English name when welsh translations are
missing.

I think the whole thing could be more coherent if we keep the Welsh
answer_settings in sync with the English answer_settings and use this
value directly when creating the FormDocument.

We would still need to handle the adding/removing of selection options
outside of the welsh translation page.