Fix rake submission email task
There is a task to allow us to directly set the submission email for a
form, without the usual code being sent or validations being performed.

When we added stricter validations to the Form model for support_email and
submission_email, the task for directly setting the submission email
needed to changed to explicitly skip validations.

The validations now run every time the form is saved. If we don't skip
the validations on saving and both the support_email and
submission_email are invalid it's impossible to the submission email.

There are 5-10 cases where both emails have leading spaces or other
issues which prevent the form being saved at all.

This commit fixes the issue by skipping validations on the form model
when setting the submission email.

Previously, we used save_draft! to save the form. This method has logic
to change the state of the form to live_with_draft if it's live or
archived_with_draft if it's archived.

For draft forms, it just calls save!.

We change the task to just use save!. When using the tasks, I don't
think we need to create a live_with_draft form, so we can just skip
that logic.