Add not null constraint for page external_id
A data migration has been run to add external_ids to all existing
pages on all environments, so it is now safe to add the not null
constraint.

Delete data_migrations.rake_spec.rb as these specs need to remove the
external_id to test the data migration, which is no longer possible.
The migration has been run on all deployed environments, but we'll
keep them around for a bit longer in case developers want to run them
on their local databases.