Remove ApiFormDocumentService and tasks
We no longer need to retrieve FormDocuments from the Forms API.

Theses tasks and supporting classes were created to aid the migration
from the Forms API.

The tasks included a task to "create or update draft FormDocuments for
all forms", which would still work. It's hard to imagine a scenario
where we would want to do this and the code is simple, so I think it's
better to remove it rather than leave it hanging around.