Add test for archived form document already exists
Add missing test case for when an archived form document already exists.

This shouldn't happen, if a form has been made live, all the archived
form documents should be deleted.

It's added because it maybe possible in the future to have multiple
archived form documents for a single form.

For example, translated forms will have multiple form documents.
'Removing' the translation from a form may mean that an archived form
exists when it shouldn't.