Add transactions to the sync methods
If there is a partial failure during the sync process, our database will
be left in an inconsistent state. This commit adds transactions to the
sync methods to ensure that the database is left in a consistent state.

To stop this, we use transactions in the sync methods.

We don't do much work in the transactions so they are short lived.

If we were to make the process more complex, we should consider an
alternative approach.