Replace FormDocumentSyncService#synchronize_form
The two places in the codebase that call this method are in
after_make_live and after_archive in the Form model.

These methods are used as callbacks for the state machine which handles
the form state.

In both cases, we already know what the form state is, so we can call
the correct method directly without needed the extra conditional.

This makes the code more direct and a bit easier to test. As the class
no longer refers to form state, we can remove the shared tests for archived and
archived_with_draft states and just test the method directly.

We rename the methods to be clearer now that they are public.