Replace strftime with as_json
When converting a date to string for JSON we use the following rails
configuration:

```app/config/application.rb
config.active_support.time_precision = 6
```

In testing, we need to fomat our timestamps in the same way. Instead of
using `strftime` we use `as_json` which will format the timestamp in the
same way as it is formatted in the JSON response.