Add support for Welsh translations in SyncService
Modify the code used to keep FormDocuments in sync with the Form's state
to account for Welsh translations.

We take care to remove old FormDocuments for languages that are no longer
used by the Form.

The archived FormDocuments must contain the same languages as the live
form. Instead of modifying the FormDocuments in place, we replace the
tag in the live FormDocuments with the archived tag.