Retrieve metrics even if form went live today
There is a bug where we are not displaying form metrics if a new
version of a form was made live today, even if it was previously live
and has metrics. We were showing the full week of metrics again a day
later.

Remove the check that prevents fetching metrics if the form was made
live today.