Add custom presenter for Welsh error summary
To make our page translation fields accessible we had to give them
custom IDs (see d8010b2a959e31a176cd621b222682edbfddbe4d).

The x-govuk form builder gem makes assumptions about the IDs of form
fields when generating the error summary, and these assumptions are
wrong because we've customised the ID. This means that the error summary
links will not point to the fields that are in error.

The gem does allow you to [configure a custom presenter for the error
summary](https://govuk-form-builder.x-govuk.org/introduction/error-handling/#custom-summary-error-presenter-injection).
This commit uses that, so we can pass the correct anchor link when
generating the validation error and it will be correctly rendered in the
error summary. If there isn't a custom URL provided, it will behave like
the default error summary.