Add first_made_live_at to forms table
We will use this column to record the timestamp of when a form is
first made live. Previously, we could only determine the date that the
form was most recently made live at by looking at the `live_at` date
in FormDocuments, but this is updated if changes to the form are made
live.