Set first_made_live_at when a form is made live for the first time
Start setting the first_made_live_at for new forms that are made live
when it isn't already set. We intend to backfill this date for
existing forms later.

Set this attribute in a `before` action for the state transition
rather than `after`, so we don't update the Form twice causing the
draft FormDocument to be re-generated twice.