Use first_made_live_at in content for displaying metrics
Rather than using the date the form was most recently made live at,
use the first_made_live_at date.

We haven't yet populated this with accurate dates for existing forms,
so fall back on using the earliest date of:
- first_made_live_at
- live_at
- the FormDocument's created_at

This will give an accurate first_made_live_at if the form has not been
archived and made live again. This isn't perfect but will display the
correct date in the metrics content for most cases. Once we have
backfilled first_made_live_at we can remove the fall-back.