Disable getting CloudWatch metrics for tests
Currently request/feature tests where we don't want to make requests
to CloudWatch rely on the `live_at` date of the FormDocument being
today. We're going to make a change to always make a request to
CloudWatch even if the form went live today, so add a setting to
explicitly disable CloudWatch requests in tests.