Change FormCopyService to use English FormDocument
The FormCopyService gets the FormDocument for the source form.

Currently the language is taken from the source form, which is always English.

The language column is being removed from the Form model.

This commit changes the FormCopyService to always use the English
FormDocument.

This will be changed when the service is updated to account for Welsh
forms.