Remove references to Form.language
We are removing the language attribute from the Form model.

Forms will model their language using the available_languages attribute,
which is an array of supported languages.

FormDocuments will have a single language.

This commit removes references to Form.language and the enum. This will
allow us to remove the language column from the Form model.