Add language to FormDocument#content
This attribute already exists on FormDocuments but it was taken from a
db column on the Form model.

Form's themselves don't have a single language. Instead they have an
array of languages.

FormDocuments contents should have a single language attribute.

This value is already on the FormDocument and API user's should know
the value, because they have added it when requesting FormDocuments.

So adding this attribute to the contents is almost redundant.

I still think it makes sense for the contents to contain the language
attribute. If it's not added here, we would need to add it in the
runner, because the value is needed when submitting a form.