Return a hash rather than array from daily_metrics_data
The service to generate a CSV from the daily metrics data will want to
return the daily totals per day in the given time range, but
CloudWatch only returns datapoints which have a non-nil value.
Therefore, returning the data as a hash keyed by the date will allow
the CSV service to get the metric value for each date in the range, if
it exists, more efficiently.

Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>