Group dependency bumps for AWS SDK gems
It appears that Amazon tends to release changes to its `aws-sdk-*` gems
in batches, and they all always depend on the `aws-sdk-core` gem, so we
tend to get several PRs to bump these gems at a time, which are all
related. So it makes sense to group dependency bumps together for these
gems.