Add validator for payment link
Adds a new validator for the GOV.UK Pay payment link format and uses it
in the payment link input.

This means we'll be able to add payment link validation to the Welsh
field in the next commit without repeating too much code in the input or
the spec files.