Rename validator test classes to avoid name collision
In two validators we were using the generic name 'Validator' in multiple places for the classes that we test custom validators on. This meant that each validator spec file would pass when run individually, but when the whole test suite was run we'd get a name collision and some of them would fail.

This commit updates our existing validator specs so that their test classes have unambiguous names derived from the name of the validator.