Make exit page validation match other markdown fields
The validation for the exit page markdown didn't include a check for
whether it was valid markdown. We have a validator for this, and we use
it on the other markdown fields, so this commit adds it to the exit page
markdown field.

The exit page heading had a maximum length of 4999 characters, which is
extremely long for a heading. Other markdown headings in the app use a
maximum of 250 characters, so this commit sets the maximum length of the
exit page heading to 250 characters as well.