Add :cy to available languages on save Welsh
When the save and continue button on Welsh translation page is used, the
available_languages attribute is not changed. This means there will be
no Welsh form docs generated for preview or live.

This commit adds :cy to the available languages attribute on save.

This happens even if the form is not marked complete to enable previews.

There currently is no way to remove a Welsh translation. This will be
added in a future PR.