Customise validation logging to avoid double counting
In the new Welsh feature, we're using nested models to encapsulate the
different fields that can be edited on one page, and merging the
validation errors so they'll all show on one error summary.

Since we're merging the validation errors, the same errors can be
present on a WelshConditionTranslationInput, a WelshPageTranslationInput
and a WelshTranslationInput at the same time.

The validation logging wasn't written to handle this, so this meant that
those errors could be logged multiple times, since each model would log
its own errors.

This commit updates the validation logic to ignore errors which are
`ActiveModel::NestedErrors`, since they should already have been logged
by the time they're nested and we want to avoid them being
double-counted in our analytics.