Add page check to validation logging logic
When looking for the form name, the validation logging looks to see if
the input has a `form` or `draft_question` field. The Welsh page and
condition translation inputs don't have direct access to `form` or
`draft_question`, but they do have access to `page`, so this commit
updates the validation logging to look for this too.