Add helper method to get number of selection options
We want to display different text on the page to configure the 'None
of the above' question depending on whether there are fewer or more
than 30 options. In the case of having more than 30 options an
autocomplete component will be shown in forms-runner for the selection
question, and the additional question for 'None of the above' will be
shown on a separate page.

This helper method will be used in the template to decide which text
to show.